requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==20.6.1
geventhttpclient==1.4.2
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
-e git://github.com/locustio/locust.git@228055e9b9bccbd8cb1a6bba429cf4a05bcdfa24#egg=locust
MarkupSafe==1.1.1
msgpack==1.0.0
psutil==5.7.0
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
six==1.15.0
urllib3==1.25.9
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
